Preliminary work on BattleshipGame
Added all of the methods + fields specified in class diagram, filled a few methods.

Added accessors + mutators for friendlyShipSunkCount and enemyShipSunkCount, since it seemed necessary.

Seems like this class may depend heavily on the methods from all the other classes. Such as when asking the name of the player, that needs to go through the GUI somehow, and we haven't determined that yet.

May also want to create a "Fleet" class to aggregate Ships, instead of using the initializeAllShips() method that I've written.