add setPosition, startPosition, endPosition, Orientation enum, checkPosition
startPosition and endPosition contain information regarding the start and end coordinates of the current ship. This allows us to check each grid position that the ship is located along when in checkPosition.
The orientation enum contains easy information on whether the ship is vertical or horizontal.

Check position will go to each position that the ship is located along and check to see if the coordinate given is within the ship.