Added placing ships and pegs
I left some code in main commented out. You can uncomment it and play around. The setPeg function now checks to see if there is a ship there, and if there is, it marks as a hit and increments its hit count.  Also added functions to print things for testing.